fix lint

+++ roomcontrol/roomcontrol_test.go
@@ -36,8 +36,8 @@ func TestRoomAddWithSavedCheck(t *testing.T) {
 
 	assert.EqualValues(t, 1, len(RoomResponse.Data), "Room number should be 1!")
 	assert.Equal(t, "TestRoomAddWithSavedCheck", RoomResponse.Data[0].Name, "Room title mismatch!")
-	assert.Equal(t, 5, RoomResponse.Data[0].Rows, "Room rows mismatch!")
-	assert.Equal(t, 7, RoomResponse.Data[0].SeatsPerRow, "Room seats mismatch!")
+	assert.Equal(t, int32(5), RoomResponse.Data[0].Rows, "Room rows mismatch!")
+	assert.Equal(t, int32(7), RoomResponse.Data[0].SeatsPerRow, "Room seats mismatch!")
 	assert.EqualValues(t, 0, RoomResponse.Data[0].Id, "id mismatch")
 }
 
@@ -89,8 +89,8 @@ func TestRoomDeleteWithSavedCheck(t *testing.T) {
 
 	assert.EqualValues(t, 1, len(RoomResponse.Data), "Room number should be 1!")
 	assert.Equal(t, "TestRoomDeleteWithSavedCheck", RoomResponse.Data[0].Name, "Room title mismatch!")
-	assert.Equal(t, 5, RoomResponse.Data[0].Rows, "Room rows mismatch!")
-	assert.Equal(t, 7, RoomResponse.Data[0].SeatsPerRow, "Room seats mismatch!")
+	assert.Equal(t, int32(5), RoomResponse.Data[0].Rows, "Room rows mismatch!")
+	assert.Equal(t, int32(7), RoomResponse.Data[0].SeatsPerRow, "Room seats mismatch!")
 	assert.EqualValues(t, 0, RoomResponse.Data[0].Id, "id mismatch")
 
 	_ = roomControl.DeleteRoom(context.TODO(), &proto.DeleteRoomRequest{Id: 0}, &response)
